requirements.txt
altair==5.4.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.16.1
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
narwhals==1.34.1
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.29.4
pyarrow==17.0.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyprojroot==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.2
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.7.0
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
zipp==3.20.2
